requirements.txt
firebase-admin
httpx
crewai
crewai[tools]
python-dotenv
fast-flights